+++ nequip/model/__init__.py
@@ -1,7 +1,7 @@
 from ._eng import EnergyModel
 from ._grads import ForceOutput
 from ._scaling import RescaleEnergyEtc, PerSpeciesRescale
-from ._weight_init import xavier_initialize_FCs
+from ._weight_init import uniform_initialize_FCs
 
 from ._build import model_from_config
 
@@ -10,6 +10,6 @@ __all__ = [
     "ForceOutput",
     "RescaleEnergyEtc",
     "PerSpeciesRescale",
-    "xavier_initialize_FCs",
+    "uniform_initialize_FCs",
     "model_from_config",
 ]