+++ nequip/model/_scaling.py
@@ -7,6 +7,9 @@ from nequip.nn import RescaleOutput, GraphModuleMixin, PerSpeciesScaleShift
 from nequip.data import AtomicDataDict, AtomicDataset
 
 
+RESCALE_THRESHOLD = 1e-6
+
+
 def RescaleEnergyEtc(
     model: GraphModuleMixin,
     config,
@@ -18,16 +21,21 @@ def RescaleEnergyEtc(
     If ``initialize`` is false, doesn't compute statistics.
     """
 
+    module_prefix = "global_rescale"
+
     global_scale = config.get(
-        "global_rescale_scale",
+        f"{module_prefix}_scale",
         f"dataset_{AtomicDataDict.FORCE_KEY}_rms"
         if AtomicDataDict.FORCE_KEY in model.irreps_out
         else f"dataset_{AtomicDataDict.TOTAL_ENERGY_KEY}_std",
     )
-    # TODO: change this default?
-    global_shift = config.get(
-        "global_rescale_shift", f"dataset_{AtomicDataDict.TOTAL_ENERGY_KEY}_mean"
-    )
+    global_shift = config.get(f"{module_prefix}_shift", None)
+
+    if global_shift is not None:
+        logging.warning(
+            f"!!!! Careful global_shift is set to {global_shift}."
+            f"The energy model will no longer be extensive"
+        )
 
     # = Get statistics of training dataset =
     if initialize:
@@ -53,11 +61,14 @@ def RescaleEnergyEtc(
         )
 
         if isinstance(global_scale, str):
+            s = global_scale
             global_scale = computed_stats[str_names.index(global_scale)]
+            logging.debug(f"Replace string {s} to {global_scale}")
         if isinstance(global_shift, str):
+            s = global_shift
             global_shift = computed_stats[str_names.index(global_shift)]
+            logging.debug(f"Replace string {s} to {global_shift}")
 
-        RESCALE_THRESHOLD = 1e-6
         if isinstance(global_scale, float) and global_scale < RESCALE_THRESHOLD:
             raise ValueError(
                 f"Global energy scaling was very low: {global_scale}. If dataset values were used, does the dataset contain insufficient variation? Maybe try disabling global scaling with global_scale=None."
@@ -66,6 +77,7 @@ def RescaleEnergyEtc(
         logging.debug(
             f"Initially outputs are globally scaled by: {global_scale}, total_energy are globally shifted by {global_shift}."
         )
+
     else:
         # Put dummy values
         if global_shift is not None:
@@ -90,12 +102,8 @@ def RescaleEnergyEtc(
             k for k in (AtomicDataDict.TOTAL_ENERGY_KEY,) if k in model.irreps_out
         ],
         shift_by=global_shift,
-        trainable_global_rescale_shift=config.get(
-            "trainable_global_rescale_shift", False
-        ),
-        trainable_global_rescale_scale=config.get(
-            "trainable_global_rescale_scale", False
-        ),
+        shift_trainable=config.get(f"{module_prefix}_shift_trainable", False),
+        scale_trainable=config.get(f"{module_prefix}_scale_trainable", False),
     )
 
 
@@ -109,47 +117,25 @@ def PerSpeciesRescale(
 
     If ``initialize`` is false, doesn't compute statistics.
     """
-    module_prefix = "PerSpeciesScaleShift_"
-
-    force_training = AtomicDataDict.FORCE_KEY in model.irreps_out
+    module_prefix = "per_species_rescale"
 
     # = Determine energy rescale type =
-    # TO DO, how to make the default consistent with the global scale function?
-    global_scale = config.get(
-        "global_rescale_scale",
+    scales = config.get(
+        module_prefix + "_scales",
         f"dataset_{AtomicDataDict.FORCE_KEY}_rms"
-        if force_training
-        else f"dataset_{AtomicDataDict.TOTAL_ENERGY_KEY}_std",
+        if AtomicDataDict.FORCE_KEY in config["train_on_keys"]
+        else f"dataset_per_atom_{AtomicDataDict.TOTAL_ENERGY_KEY}_std",
+    )
+    shifts = config.get(
+        module_prefix + "_shifts",
+        f"dataset_per_atom_{AtomicDataDict.TOTAL_ENERGY_KEY}_mean",
     )
 
-    # TODO: how to make the default consistent with rescale?
-    global_shift = config.get(
-        "global_rescale_shift", f"dataset_{AtomicDataDict.TOTAL_ENERGY_KEY}_mean"
-    )
-    scales = config.get(module_prefix + "scales", None)
-    shifts = config.get(module_prefix + "shifts", None)
-    trainable = config.get(module_prefix + "trainable", False)
-    kwargs = config.get(module_prefix + "kwargs", {})
-
-    if global_shift is not None:
-        if trainable or not (scales is None and shifts is None):
-            logging.warning(
-                f"!!!! Careful global_shift is set to {global_shift}."
-                f"This is not a good set up with per species shifts: {shifts}"
-                f"and scales: {scales} that are trainable={trainable}"
-            )
-    if not trainable:
-        if scales is None and shifts is None:
-            return model
-        elif scales == 1.0 and shifts == 0.0:
-            return model
-
-    logging.info(f"Enable per species scale/shift")
-
-    # = Determine what statistics need to be compute =
+    # = Determine what statistics need to be compute =\
+    arguments_in_dataset_units = None
     if initialize:
         str_names = []
-        for value in [scales, shifts, global_scale]:
+        for value in [scales, shifts]:
             if isinstance(value, str):
                 str_names += [value]
             elif (
@@ -161,50 +147,71 @@ def PerSpeciesRescale(
                 # valid values
                 pass
             else:
-                raise ValueError(f"Invalid value `{value}`")
+                raise ValueError(f"Invalid value `{value}` of type {type(value)}")
+
+        if len(str_names) == 2:
+            # Both computed from dataset
+            arguments_in_dataset_units = True
+        elif len(str_names) == 1:
+            assert config[
+                module_prefix + "arguments_in_dataset_units"
+            ], "Requested to set either the shifts or scales of the per_species_rescale using dataset values, but chose to provide the other in non-dataset units. Please give the explictly specified shifts/scales in dataset units and set per_species_rescale_arguments_in_dataset_units"
 
         # = Compute shifts and scales =
         computed_stats = _compute_stats(
             str_names=str_names,
             dataset=dataset,
             stride=config.dataset_statistics_stride,
-            kwargs=kwargs,
+            kwargs=config.get(module_prefix + "_kwargs", {}),
         )
 
         if isinstance(scales, str):
+            s = scales
             scales = computed_stats[str_names.index(scales)]
+            logging.debug(f"Replace string {s} to {scales}")
+        elif isinstance(scales, list):
+            scales = torch.as_tensor(scales)
 
         if isinstance(shifts, str):
+            s = shifts
             shifts = computed_stats[str_names.index(shifts)]
+            logging.debug(f"Replace string {s} to {shifts}")
+        elif isinstance(shifts, list):
+            shifts = torch.as_tensor(shifts)
 
-        if isinstance(global_scale, str):
-            global_scale = computed_stats[str_names.index(global_scale)]
-
-        if global_scale is not None:
-            if scales is not None:
-                scales = scales / global_scale
-            if shifts is not None:
-                shifts = shifts / global_scale
+        if scales.min() < RESCALE_THRESHOLD:
+            raise ValueError(
+                f"Per species energy scaling was very low: {scales}. Maybe try setting {module_prefix}_scales = 1."
+            )
 
     else:
+
         # Put dummy values
-        scales = None
-        shifts = None
+        # the real ones will be loaded from the state dict later
+        # note that the state dict includes buffers,
+        # so this is fine regardless of whether its trainable.
+        scales = 1.0
+        shifts = 0.0
+        # values correctly scaled according to where the come from
+        # will be brought from the state dict later,
+        # so what you set this to doesnt matter:
+        arguments_in_dataset_units = False
 
     # insert in per species shift
+    params = dict(
+        field=AtomicDataDict.PER_ATOM_ENERGY_KEY,
+        out_field=AtomicDataDict.PER_ATOM_ENERGY_KEY,
+        shifts=shifts,
+        scales=scales,
+    )
+    if arguments_in_dataset_units is not None:
+        params["arguments_in_dataset_units"] = arguments_in_dataset_units
     model.insert_from_parameters(
         before="total_energy_sum",
-        name="per_species_scale_shift",
+        name=module_prefix,
         shared_params=config,
         builder=PerSpeciesScaleShift,
-        params=dict(
-            field=AtomicDataDict.PER_ATOM_ENERGY_KEY,
-            out_field=AtomicDataDict.PER_ATOM_ENERGY_KEY,
-            num_types=config.num_types,
-            shifts=shifts,
-            scales=scales,
-            trainable=trainable,
-        ),
+        params=params,
     )
 
     logging.debug(f"Atomic outputs are scaled by: {scales}, shifted by {shifts}.")