+++ nequip/model/_weight_init.py
@@ -31,31 +31,22 @@ def unit_uniform_init_(t: torch.Tensor):
 
 
 # TODO: does this normalization make any sense
-def unit_orthogonal_init_(t: torch.Tensor):
-    """Orthogonal init with <x_i^2> = 1"""
-    assert t.ndim == 2
-    torch.nn.init.orthogonal_(t, gain=math.sqrt(max(t.shape)))
+# def unit_orthogonal_init_(t: torch.Tensor):
+#     """Orthogonal init with <x_i^2> = 1"""
+#     assert t.ndim == 2
+#     torch.nn.init.orthogonal_(t, gain=math.sqrt(max(t.shape)))
 
 
-# TODO: more inits
-
-
-def _xavier_initialize_fcs(mod: torch.nn.Module):
-    """Initialize ``e3nn.nn.FullyConnectedNet``s and ``torch.nn.Linear``s with Xavier uniform initialization"""
-    if isinstance(mod, e3nn.nn.FullyConnectedNet):
-        for layer in mod:
-            # in FC:
-            # h_in, _h_out = W.shape
-            # W = W / h_in**0.5
-            torch.nn.init.xavier_uniform_(
-                layer.weight, gain=layer.weight.shape[0] ** 0.5
-            )
-    elif isinstance(mod, torch.nn.Linear):
-        torch.nn.init.xavier_uniform_(mod.weight)
-
-
-def xavier_initialize_FCs(model: GraphModuleMixin, initialize: bool):
+def uniform_initialize_FCs(model: GraphModuleMixin, initialize: bool):
+    """Initialize ``e3nn.nn.FullyConnectedNet``s with unit uniform initialization"""
     if initialize:
+
+        def _uniform_initialize_fcs(mod: torch.nn.Module):
+            if isinstance(mod, e3nn.nn.FullyConnectedNet):
+                for layer in mod:
+                    # in FC, normalization is expected
+                    unit_uniform_init_(layer.weight)
+
         with torch.no_grad():
-            model.apply(_xavier_initialize_fcs)
+            model.apply(_uniform_initialize_fcs)
     return model