+++ tests/unit/utils/test_weight_init.py
@@ -2,10 +2,10 @@ import pytest
 
 import torch
 
-from nequip.model._weight_init import unit_uniform_init_, unit_orthogonal_init_
+from nequip.model._weight_init import unit_uniform_init_
 
 
-@pytest.mark.parametrize("init_func_", [unit_uniform_init_, unit_orthogonal_init_])
+@pytest.mark.parametrize("init_func_", [unit_uniform_init_])
 def test_2mom(init_func_):
     t = torch.empty(1000, 100)
     init_func_(t)